using defer to guarantee that status is updated at the end of reconcile.
We would like to ensure the reconcile loop always updates the CR's
status. As it was written prior, this required the developer to remember
not to modify or return before running that logic in the reconcile
loop. With the use of defer, this requirement is now explicitly
expressed in the code. As reconcile returns, the function defined within
the defer statement is called and will update the status. This will run
no matter when reconcile returns.