Adding readiness probe sidecar.
As an OCM add-on, the ocs-osd-operator needs a way to communicate its
status to OCM. OCM is currently limited to using the CSV status of the
add-on, which it gets every 8 minutes through OLM telemetry. In order to
control when the operator is reported as installed, a readiness probe
was added, along with logic to only report the operator as ready when
all of its components are in their desired state.

Signed-off-by: Renan Campos <rcampos@redhat.com>