Updating functionality of deploy-with-olm.
For the operator to run without error, certain objects must be present
on the cluster prior to deployment:
- namespace/openshift-storage
- secret/add-${ADDON_NAME}-parameters
- secret/${ADDON_NAME}-pagerduty
- secret/${ADDON_NAME}-deadmanssnitch

The CSV of the bundle must also have two environment variables set for
it to run properly:
- ADDON_NAME
- SOP_ENDPOINT

This update ensures these objects are present and provides a way for
the user to customize their parameters by defining them as an
environment variable prior to running.

The bundle generated by this script will be in ./tools/bundle, so its
changes do not disrupt ./bundle, which is intended to be used by the
developer.

Signed-off-by: Renan Campos <rcampos@redhat.com>