Created a Dummy commit
This Dummy commit is created because the image which was built previously was incorrect. Whenever there is change in Dockerfile of this repo, there might be a change in ocs-deployer dist-git repo(brew build) Dockerfile as well. Once an image is tagged with a specific hash, we can't create new image with same hash. So a new dummy commit is created which will give us new hash to create the new image.

Signed-off-by: bindrad <dbindra@redhat.com>