Using AWS data for network policy setup.
Provider clusters run storage pods on the host network namespace of
their worker nodes. To properly set up the network policy, the machine
CIDR must be known. In AWS, this can be determined by making a call to
the IMDSv2 server that is available in any VPC. This PR creates a
deployment to gather the needed data, and modifies the reconcilation
loop to make use of the AWS data when setting up the network policies.

Signed-off-by: Renan Campos <rcampos@redhat.com>